refacto(design): more restructuring * moved design specific actions to src/design/actions * removed obsolete code * various related cleanups
